Abstract out component overlaying
